Remove Extras, migrate mapK, Update Docs
