--- cats/extras/shared/src/main/scala/io/chrisdavenport/log4cats/extras/implicits.scala
@@ -1,5 +1,0 @@
-package io.chrisdavenport.log4cats.extras
-
-import io.chrisdavenport.log4cats.extras.syntax.loggersyntax
-
-object implicits extends loggersyntax
