+++ core/shared/src/main/scala/io/chrisdavenport/log4cats/SelfAwareStructuredLogger.scala
@@ -1,6 +1,11 @@
 package io.chrisdavenport.log4cats
 
-trait SelfAwareStructuredLogger[F[_]] extends SelfAwareLogger[F] with StructuredLogger[F]
+import cats._
+
+trait SelfAwareStructuredLogger[F[_]] extends SelfAwareLogger[F] with StructuredLogger[F] {
+  override def mapK[G[_]](fk: F ~> G): SelfAwareStructuredLogger[G] = 
+    SelfAwareStructuredLogger.mapK(fk)(this)
+}
 
 object SelfAwareStructuredLogger {
   def apply[F[_]](implicit ev: SelfAwareStructuredLogger[F]): SelfAwareStructuredLogger[F] = ev
@@ -58,4 +63,64 @@ object SelfAwareStructuredLogger {
     def trace(ctx: Map[String, String], t: Throwable)(message: => String): F[Unit] =
       sl.trace(outer ++ ctx, t)(message)
   }
+
+
+  private def mapK[G[_], F[_]](
+      f: G ~> F
+  )(logger: SelfAwareStructuredLogger[G]): SelfAwareStructuredLogger[F] =
+    new SelfAwareStructuredLogger[F] {
+      def isTraceEnabled: F[Boolean] =
+        f(logger.isTraceEnabled)
+      def isDebugEnabled: F[Boolean] =
+        f(logger.isDebugEnabled)
+      def isInfoEnabled: F[Boolean] =
+        f(logger.isInfoEnabled)
+      def isWarnEnabled: F[Boolean] =
+        f(logger.isWarnEnabled)
+      def isErrorEnabled: F[Boolean] =
+        f(logger.isErrorEnabled)
+
+      def trace(ctx: Map[String, String])(msg: => String): F[Unit] =
+        f(logger.trace(ctx)(msg))
+      def debug(ctx: Map[String, String])(msg: => String): F[Unit] =
+        f(logger.debug(ctx)(msg))
+      def info(ctx: Map[String, String])(msg: => String): F[Unit] =
+        f(logger.info(ctx)(msg))
+      def warn(ctx: Map[String, String])(msg: => String): F[Unit] =
+        f(logger.warn(ctx)(msg))
+      def error(ctx: Map[String, String])(msg: => String): F[Unit] =
+        f(logger.error(ctx)(msg))
+
+      def error(t: Throwable)(message: => String): F[Unit] =
+        f(logger.error(t)(message))
+      def warn(t: Throwable)(message: => String): F[Unit] =
+        f(logger.warn(t)(message))
+      def info(t: Throwable)(message: => String): F[Unit] =
+        f(logger.info(t)(message))
+      def debug(t: Throwable)(message: => String): F[Unit] =
+        f(logger.debug(t)(message))
+      def trace(t: Throwable)(message: => String): F[Unit] =
+        f(logger.trace(t)(message))
+      def error(message: => String): F[Unit] =
+        f(logger.error(message))
+      def warn(message: => String): F[Unit] =
+        f(logger.warn(message))
+      def info(message: => String): F[Unit] =
+        f(logger.info(message))
+      def debug(message: => String): F[Unit] =
+        f(logger.debug(message))
+      def trace(message: => String): F[Unit] =
+        f(logger.trace(message))
+
+      def trace(ctx: Map[String, String], t: Throwable)(msg: => String): F[Unit] =
+        f(logger.trace(ctx, t)(msg))
+      def debug(ctx: Map[String, String], t: Throwable)(msg: => String): F[Unit] =
+        f(logger.debug(ctx, t)(msg))
+      def info(ctx: Map[String, String], t: Throwable)(msg: => String): F[Unit] =
+        f(logger.info(ctx, t)(msg))
+      def warn(ctx: Map[String, String], t: Throwable)(msg: => String): F[Unit] =
+        f(logger.warn(ctx, t)(msg))
+      def error(ctx: Map[String, String], t: Throwable)(msg: => String): F[Unit] =
+        f(logger.error(ctx, t)(msg))
+    }
 }
