+++ core/shared/src/main/scala/io/chrisdavenport/log4cats/StructuredLogger.scala
@@ -1,5 +1,7 @@
 package io.chrisdavenport.log4cats
 
+import cats._
+
 trait StructuredLogger[F[_]] extends Logger[F] {
   def trace(ctx: Map[String, String])(msg: => String): F[Unit]
   def trace(ctx: Map[String, String], t: Throwable)(msg: => String): F[Unit]
@@ -11,6 +13,8 @@ trait StructuredLogger[F[_]] extends Logger[F] {
   def warn(ctx: Map[String, String], t: Throwable)(msg: => String): F[Unit]
   def error(ctx: Map[String, String])(msg: => String): F[Unit]
   def error(ctx: Map[String, String], t: Throwable)(msg: => String): F[Unit]
+  override def mapK[G[_]](fk: F ~> G): StructuredLogger[G] = 
+    StructuredLogger.mapK(fk)(this)
 }
 
 object StructuredLogger {
@@ -61,4 +65,51 @@ object StructuredLogger {
     def trace(ctx: Map[String, String], t: Throwable)(message: => String): F[Unit] =
       sl.trace(outer ++ ctx, t)(message)
   }
+
+
+  private def mapK[G[_], F[_]](f: G ~> F)(logger: StructuredLogger[G]): StructuredLogger[F] =
+    new StructuredLogger[F] {
+      def trace(ctx: Map[String, String])(msg: => String): F[Unit] =
+        f(logger.trace(ctx)(msg))
+      def debug(ctx: Map[String, String])(msg: => String): F[Unit] =
+        f(logger.debug(ctx)(msg))
+      def info(ctx: Map[String, String])(msg: => String): F[Unit] =
+        f(logger.info(ctx)(msg))
+      def warn(ctx: Map[String, String])(msg: => String): F[Unit] =
+        f(logger.warn(ctx)(msg))
+      def error(ctx: Map[String, String])(msg: => String): F[Unit] =
+        f(logger.error(ctx)(msg))
+
+      def error(t: Throwable)(message: => String): F[Unit] =
+        f(logger.error(t)(message))
+      def warn(t: Throwable)(message: => String): F[Unit] =
+        f(logger.warn(t)(message))
+      def info(t: Throwable)(message: => String): F[Unit] =
+        f(logger.info(t)(message))
+      def debug(t: Throwable)(message: => String): F[Unit] =
+        f(logger.debug(t)(message))
+      def trace(t: Throwable)(message: => String): F[Unit] =
+        f(logger.trace(t)(message))
+      def error(message: => String): F[Unit] =
+        f(logger.error(message))
+      def warn(message: => String): F[Unit] =
+        f(logger.warn(message))
+      def info(message: => String): F[Unit] =
+        f(logger.info(message))
+      def debug(message: => String): F[Unit] =
+        f(logger.debug(message))
+      def trace(message: => String): F[Unit] =
+        f(logger.trace(message))
+
+      def trace(ctx: Map[String, String], t: Throwable)(msg: => String): F[Unit] =
+        f(logger.trace(ctx, t)(msg))
+      def debug(ctx: Map[String, String], t: Throwable)(msg: => String): F[Unit] =
+        f(logger.debug(ctx, t)(msg))
+      def info(ctx: Map[String, String], t: Throwable)(msg: => String): F[Unit] =
+        f(logger.info(ctx, t)(msg))
+      def warn(ctx: Map[String, String], t: Throwable)(msg: => String): F[Unit] =
+        f(logger.warn(ctx, t)(msg))
+      def error(ctx: Map[String, String], t: Throwable)(msg: => String): F[Unit] =
+        f(logger.error(ctx, t)(msg))
+    }
 }
