+++ core/shared/src/main/scala/io/chrisdavenport/log4cats/extras/LogMessage.scala
