+++ core/shared/src/main/scala/io/chrisdavenport/log4cats/extras/WriterLogger.scala
@@ -68,29 +68,29 @@ object WriterLogger {
     }
   }
 
-  def run[F[_]: Logger: Applicative, G[_]: Foldable]: Writer[G[LogMessage], ?] ~> F =
+  def run[F[_]: Applicative, G[_]: Foldable](l: Logger[F]): Writer[G[LogMessage], ?] ~> F =
     new ~>[Writer[G[LogMessage], ?], F] {
       def logMessage(logMessage: LogMessage): F[Unit] = logMessage match {
         case LogMessage(LogLevel.Error, Some(t), m) =>
-          Logger[F].error(t)(m)
+          l.error(t)(m)
         case LogMessage(LogLevel.Error, None, m) =>
-          Logger[F].error(m)
+          l.error(m)
         case LogMessage(LogLevel.Warn, Some(t), m) =>
-          Logger[F].warn(t)(m)
+          l.warn(t)(m)
         case LogMessage(LogLevel.Warn, None, m) =>
-          Logger[F].warn(m)
+          l.warn(m)
         case LogMessage(LogLevel.Info, Some(t), m) =>
-          Logger[F].info(t)(m)
+          l.info(t)(m)
         case LogMessage(LogLevel.Info, None, m) =>
-          Logger[F].info(m)
+          l.info(m)
         case LogMessage(LogLevel.Debug, Some(t), m) =>
-          Logger[F].debug(t)(m)
+          l.debug(t)(m)
         case LogMessage(LogLevel.Debug, None, m) =>
-          Logger[F].debug(m)
+          l.debug(m)
         case LogMessage(LogLevel.Trace, Some(t), m) =>
-          Logger[F].trace(t)(m)
+          l.trace(t)(m)
         case LogMessage(LogLevel.Trace, None, m) =>
-          Logger[F].trace(m)
+          l.trace(m)
       }
 
       def apply[A](fa: Writer[G[LogMessage], A]): F[A] = {
