+++ noop/shared/src/main/scala/io/chrisdavenport/log4cats/noop/NoOpLogger.scala
