+++ slf4j/src/main/scala/io/chrisdavenport/log4cats/slf4j/internal/LoggerMacros.scala
