+++ slf4j/src/test/scala/io/chrisdavenport/log4cats/slf4j/Slf4jLoggerMacroCompilationTests.scala
