+++ testing/shared/src/main/scala/io/chrisdavenport/log4cats/testing/TestingLogger.scala
